Renamed appleseed module to appleseed_maya.
