+++ module/scripts/appleseed_maya/aenodetemplates.py
