+++ module/scripts/appleseed_maya/aetemplate/aeasdisneymaterialtemplate.py
