+++ module/scripts/appleseed_maya/aetemplate/aeaslayeredshadertemplate.py
