+++ module/scripts/appleseed_maya/aetemplate/aemtap_standinlocatortemplate.py
