+++ module/scripts/appleseed_maya/aetemplate/aemtap_standinmeshnodetemplate.py
