+++ module/scripts/appleseed_maya/binmeshtranslator.py
