+++ module/scripts/appleseed_maya/commonglobals.py
