+++ module/scripts/appleseed_maya/metashadernode.py
