+++ module/scripts/appleseed_maya/optimizetextures.py
