+++ module/scripts/appleseed_maya/osltools.py
