+++ module/scripts/appleseed_maya/plug-ins/loadshadersplugin.py
@@ -27,11 +27,12 @@
 #
 
 import maya.OpenMayaMPx as OpenMayaMPx
-import appleseed.shadingnodes as shadingNodes
+import appleseed_maya.shadingnodes as shadingNodes
 
 def initializePlugin(mobject):
     mplugin = OpenMayaMPx.MFnPlugin(mobject)
     shadingNodes.registerShadingNodes(mplugin)
+
 
 def uninitializePlugin(mobject):
     mplugin = OpenMayaMPx.MFnPlugin(mobject)
