+++ module/scripts/appleseed_maya/shadingnodes.py
