+++ src/appleseedswatchrenderer.cpp
@@ -93,7 +93,7 @@ AppleseedSwatchRenderer::AppleseedSwatchRenderer()
     }
 
     MStringArray oslDirs;
-    MGlobal::executePythonCommand("import appleseed.osltools as osl; osl.getOSODirs();", oslDirs, false, false);
+    MGlobal::executePythonCommand("import appleseed_maya.osltools as osl; osl.getOSODirs();", oslDirs, false, false);
 
     for (unsigned int i = 0; i < oslDirs.length(); i++)
         mProject->search_paths().push_back(oslDirs[i].asChar());
